💄 put technologies in a card and finish style
